--- .gitmodules
@@ -1,4 +1,0 @@
-[submodule "build_utils"]
-	path = build_utils
-	url = git@github.com:rbkmoney/build_utils.git
-	branch = master
