--- src/main/java/com/rbkmoney/http/client/domain/Response.java
@@ -1,14 +1,0 @@
-package com.rbkmoney.http.client.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Response<T> {
-
-    private T entity;
-
-}
