--- src/main/java/com/rbkmoney/http/client/exception/RemoteInvocationException.java
@@ -1,22 +1,0 @@
-package com.rbkmoney.http.client.exception;
-
-public class RemoteInvocationException extends RuntimeException {
-    public RemoteInvocationException() {
-    }
-
-    public RemoteInvocationException(String message) {
-        super(message);
-    }
-
-    public RemoteInvocationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public RemoteInvocationException(Throwable cause) {
-        super(cause);
-    }
-
-    public RemoteInvocationException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
