--- src/main/java/com/rbkmoney/http/client/pool/HttpClientPool.java
@@ -1,9 +1,0 @@
-package com.rbkmoney.http.client.pool;
-
-public interface HttpClientPool<T, R> {
-
-    R get(T config);
-
-    void destroy();
-
-}
