--- src/main/java/com/rbkmoney/http/client/properties/KeyStoreProperties.java
@@ -1,15 +1,0 @@
-package com.rbkmoney.http.client.properties;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class KeyStoreProperties {
-
-    private String certificateFolder;
-    private String type;
-    private String password;
-
-}
