--- src/main/java/com/rbkmoney/http/client/properties/SslRequestConfig.java
@@ -1,14 +1,0 @@
-package com.rbkmoney.http.client.properties;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class SslRequestConfig {
-
-    private String certFileName;
-    private String certType;
-    private String certPass;
-
-}
