+++ src/main/java/dev/vality/http/client/domain/Response.java
@@ -1,0 +1,14 @@
+package dev.vality.http.client.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Response<T> {
+
+    private T entity;
+
+}
