+++ src/main/java/dev/vality/http/client/pool/HttpClientPool.java
@@ -1,0 +1,9 @@
+package dev.vality.http.client.pool;
+
+public interface HttpClientPool<T, R> {
+
+    R get(T config);
+
+    void destroy();
+
+}
