+++ src/main/java/dev/vality/http/client/properties/SslRequestConfig.java
@@ -1,0 +1,14 @@
+package dev.vality.http.client.properties;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class SslRequestConfig {
+
+    private String certFileName;
+    private String certType;
+    private String certPass;
+
+}
